# internal/webui/openapi.yaml
openapi: 3.0.3
info:
  title: gourl API
  description: |
    Lightweight self-hosted URL shortener API.

    Admin endpoints accept either a session cookie (from
    `POST /api/v1/auth/login`) or `Authorization: Bearer <token>` (tokens are
    created in Settings). With no admin password configured the server runs
    in setup mode: admin endpoints refuse with `403 setup_required` (bearer
    tokens still authenticate) until the setup flow (`POST /api/v1/auth/setup`)
    sets the first password. The redirect route `GET /{code}` is always
    public and never requires auth.
  version: "0.1.0"
servers:
  - url: /api/v1
...
paths:
...
  /links:
    get:
      tags: [links]
      summary: List links (paginated, searchable, sortable, expiry-filterable)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: q, in: query, schema: { type: string }, description: Match on code/url/title }
        - { name: expires, in: query, schema: { type: string, enum: ["", active, expired] }, description: "Filter by expiry state (default: all)" }
        - { name: sort, in: query, schema: { type: string, enum: [created_at, clicks, code] } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SetupRequired" }
        "200":
          description: Link list with total count
          content:
            application/json:
              schema:
                type: object
                properties:
                  links: { type: array, items: { $ref: "#/components/schemas/Link" } }
                  total: { type: integer }
                  page: { type: integer }
                  page_size: { type: integer }
    post:
      tags: [links]
      summary: Create a link (title/description are fetched in the background)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLink" }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SetupRequired" }
        "201": { description: Created, returns the link with full short URLs }
        "400": { description: Invalid url/code or reserved code }
        "409": { description: Code already in use }
    delete:
      tags: [links]
      summary: Delete multiple links at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [codes]
              properties:
                codes:
                  type: array
                  maxItems: 500
                  items: { type: string }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SetupRequired" }
        "200":
          description: Number of links actually deleted (absent codes are skipped). Daily click history is kept, as always.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: integer }
...
  /links/{code}:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string }, description: "Short code (may contain /, ASCII or Chinese)" }
    get:
      tags: [links]
      summary: Get a link
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SetupRequired" }
        "200": { description: Link }
        "404": { description: Not found }
    patch:
      tags: [links]
      summary: Update a link (url change re-fetches title in the background unless overridden)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                url: { type: string }
                title: { type: string }
                description: { type: string }
                expires_at: { type: integer }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SetupRequired" }
        "200": { description: Updated link }
        "404": { description: Not found }
        "409": { description: New code already in use }
    delete:
      tags: [links]
      summary: Delete a link (its daily click records are deliberately kept for history)
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/SetupRequired" }
        "204": { description: Deleted }
        "404": { description: Not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Missing or invalid session cookie / bearer token
    SetupRequired:
      description: "setup_required — no admin password configured yet, run the setup flow first (bearer tokens still authenticate)"
  schemas:
    Link:
      type: object
      properties:
        code: { type: string }
        url: { type: string }
        title: { type: string }
        description: { type: string }
        expires_at: { type: integer }
        click_count: { type: integer }
        created_at: { type: integer }
        updated_at: { type: integer }
        urls:
          type: array
          description: Full short URLs (primary + extra base URLs)
          items: { type: string }
    CreateLink:
      type: object
      required: [url]
      properties:
        url: { type: string, description: "Absolute http(s) URL" }
        code: { type: string, description: "Optional custom code, may contain / (≤5 levels, ASCII or Chinese); empty = random" }
        title: { type: string }
        description: { type: string }
        expires_at:
          oneOf:
            - type: integer
              description: Unix seconds; 0 = never
            - type: string
              description: "yyyy-MM-dd calendar date, parsed at local midnight"
        click_count:
          type: integer
          description: "Import-time override (batch only); single creates ignore it"
        created_at:
          type: integer
          description: "Import-time override (batch only); single creates ignore it"
...
security:
  - bearerAuth: []